+++ tokogae-data/tokogae-data-event/tokogae-data-event-api/src/main/java/com/tokogae/data/event/service/FoodItemLocalServiceUtil.java
@@ -54,7 +54,7 @@ public class FoodItemLocalServiceUtil {
 	public static FoodItem addFoodItem(
 			long userId, long subjectId, long occurDay, int occurDaySegment,
 			long occurTime, String name, String locationOfOrigin, String brand,
-			int quantity, String quantityUnit)
+			double quantity, String quantityUnit)
 		throws PortalException {
 
 		return getService().addFoodItem(
@@ -294,7 +294,7 @@ public class FoodItemLocalServiceUtil {
 
 	public static FoodItem updateFoodItem(
 			long foodItemId, long occurDay, int occurDaySegment, long occurTime,
-			String name, String locationOfOrigin, String brand, int quantity,
+			String name, String locationOfOrigin, String brand, double quantity,
 			String quantityUnit)
 		throws PortalException {
